refactor(monorepo): phase 4 — extract route handler into core
Move the 1323-line badge request handler from web route into
@shieldcn/core/route-handler as reusable handleBadgeGET/handleBadgePUT.

- fetchBadgeData (provider routing, 700+ lines)
- getDefaultLogoSlug (icon mapping)
- parseFormat, parseGradient, isLightHex, brandedFg helpers
- Full icon/color/variant resolution pipeline
- PNG rendering via @resvg/resvg-wasm
- Optional onTrack callback for analytics (web passes trackEvent)

Web route handler reduced from 1323 → 26 lines.
Engine can now import the same handler without analytics.

Verified:
- next build succeeds
- SVG, PNG, JSON, shields.json all work
- Branded, outline, gradient variants render correctly
- Error badges render
- Memo PUT works (returns DB error as expected without Postgres)

Part of #43, closes #48
